+++ src/main/kotlin/server/endpoints/create/NewZoneEndpoint.kt
@@ -14,8 +14,9 @@ import io.ktor.server.routing.RoutingContext
 import java.io.File
 import java.net.URI
 import server.endpoints.SecureEndpointBase
-import server.error.Errors
+import server.error.Error
 import server.error.Errors.MissingData
+import server.error.Errors.ParentNotFound
 import server.request.save
 import server.response.respondFailure
 import server.response.respondSuccess
@@ -37,6 +38,7 @@ object NewZoneEndpoint : SecureEndpointBase("/zone") {
         var imageFile: File? = null
         var kmzFile: File? = null
 
+        var error: Error? = null
         receiveMultipart(
             forEachFormItem = { partData ->
                 when (partData.name) {
@@ -46,7 +48,7 @@ object NewZoneEndpoint : SecureEndpointBase("/zone") {
                     "points" -> points = Json.decodeFromString(partData.value)
                     "area" -> ServerDatabase.instance.query {
                         area = Area.findById(partData.value.toInt())
-                            ?: return@query respondFailure(Errors.ParentNotFound)
+                            ?: return@query ParentNotFound.let { error = it }
                     }
                 }
             },
@@ -57,6 +59,10 @@ object NewZoneEndpoint : SecureEndpointBase("/zone") {
                 }
             }
         )
+        if (error != null) {
+            call.respondFailure(error)
+            return
+        }
 
         if (isAnyNull(displayName, webUrl, imageFile, kmzFile, area)) {
             imageFile?.delete()
